% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mvnpdf.R, R/mvnpdfoptim.R,
%   R/mvnpdfsmart.R
\name{mvnpdfC}
\alias{mvnpdfC}
\alias{mvnpdf}
\alias{mvnpdfoptim}
\alias{mvnpdfsmart}
\alias{mvnpdf_invC}
\title{Based on the implementation from Nino Hardt and Dicko Ahmadou
http://gallery.rcpp.org/articles/dmvnorm_arma/
(accessed in August 2014)}
\usage{
mvnpdfC(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdf_invC(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfC(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{a matrix, with \code{n} columns \code{the observations} and
\code{p} rows}

\item{mean}{a vector of means}

\item{varcovM}{a variance-covariance matrix}

\item{Log}{a logical parameter, with default value to \code{TRUE}.}
}
\value{
a list containing the input matrix x and y the multivariate-normal
}
\description{
We want to compute the value of the density of a
multivariate normal distribution on R p at n points.
Our function must be applicable for any multivariate
normal distribution (i.e. any mean vector in R p and
variance-covariance matrix of any order of p ), and
we wish to compute all the values of the density evaluated
at the n points in a single call of the function.
}
\examples{

if(interactive()){
k <- 50
p <- 50
varcovM <- matrix(NA, p, k)
x <- matrix(NA, p, k)
for(k in 1:k) {
varcovM[,k] <- abs(rnorm(p,.5,1))
x[,k] <- rnorm(p,.5,1)
}
mean <- as.array(rnorm(ncol(varcovM), .5, 1))
mvnpdf(x = x,mean = mean, varcovM = varcovM, Log=FALSE)}

}
